+++ apps/docs/src/app/app.module.ts
@@ -5,6 +5,7 @@ import { MarkdownModule } from 'ngx-markdown';
 
 import { AppComponent } from './app.component';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { ContentDensityService } from '@fundamental-ngx/core';
 
 const routes: Routes = [
     {
@@ -27,6 +28,9 @@ const routes: Routes = [
         RouterModule.forRoot(routes, { useHash: true, relativeLinkResolution: 'legacy' }),
         MarkdownModule.forRoot({ loader: HttpClient })
     ],
-    bootstrap: [AppComponent]
+    bootstrap: [AppComponent],
+    providers: [
+        ContentDensityService
+    ]
 })
 export class AppModule {}