+++ libs/core/src/lib/wizard/wizard-step-indicator/wizard-step-indicator.component.ts
@@ -4,18 +4,23 @@ import {
     Component,
     EventEmitter,
     Input,
+    OnDestroy,
+    OnInit,
+    Optional,
     Output,
     ViewChild
 } from '@angular/core';
 import { WizardStepComponent } from '../wizard-step/wizard-step.component';
 import { ActionSheetComponent } from '../../action-sheet/action-sheet.component';
+import { ContentDensityService } from '../../utils/public_api';
+import { Subscription } from 'rxjs';
 
 @Component({
     selector: 'fd-wizard-step-indicator',
     templateUrl: './wizard-step-indicator.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class WizardStepIndicatorComponent {
+export class WizardStepIndicatorComponent implements OnInit, OnDestroy {
     /**
      * The icon to use for this step.
      */
@@ -26,7 +31,7 @@ export class WizardStepIndicatorComponent {
      * Whether or not the step indicator (specifically, the action sheet) is compact.
      */
     @Input()
-    compact = false;
+    compact?: boolean;
 
     /**
      * Event emitted when this step indicator is clicked.
@@ -40,7 +45,25 @@ export class WizardStepIndicatorComponent {
     /** @hidden */
     stackedItems: WizardStepComponent[];
 
-    constructor(private _cdRef: ChangeDetectorRef) {}
+    /** @hidden */
+    private _subscriptions = new Subscription();
+
+    constructor(private _cdRef: ChangeDetectorRef, @Optional() private _contentDensityService: ContentDensityService) {}
+
+    /** @hidden */
+    ngOnInit(): void {
+        if (this.compact === undefined && this._contentDensityService) {
+            this._subscriptions.add(this._contentDensityService._contentDensityListener.subscribe(density => {
+                this.compact = density !== 'cozy';
+                this._cdRef.markForCheck();
+            }))
+        }
+    }
+
+    /** @hidden */
+    ngOnDestroy(): void {
+        this._subscriptions.unsubscribe();
+    }
 
     /** @hidden */
     stepItemClicked(step?: WizardStepComponent, event?: MouseEvent): void {